Add files via upload
Frequency analysis attack is done by using the knowledge of the most used character in English alphabet (or the language used for the plaintext message). This is done by analyzing the ciphertext to find most frequent character and then subtracting its position from the most frequent character in the language used for plaintext. This will provide the key and that key can be used to decrypt the ciphertext into plaintext.