Add files via upload
This code breaks Caesar Cipher using Grover's algorithm. The circuit is picked up from https://eprint.iacr.org/2021/554.pdf . The Oracle holds the plaintext and we query and try to solve reducing the number of queries compared to Classical brute force method, which needs 'N' queries in worst case, where N is the size of the alphabet set.