Add files via upload
This code provides the Brute-Force attack on Caesar Cipher. The idea is to try all 26 keys (if we use English alphabet then there are only 26 alphabets hence 26 possible keys). As the search-space is very small we can try all possible keys on ciphertext and decrypt. This is called "Brute-Force" attack.